Auto Reset after 2h 5mins
This is so if the bot gets stuck anywhere, like with the helix use
thing, it will reset after 2h and 5 mins meaning that it will reset and
continue with runs, and not be stuck at places for a while! it should
not effect the speedrun as we are aiming for times under 1:55 anyways!